[#154036434] Implement Splash screen and start using correct package names.
