+++ app/src/main/java/com/blackcrowsys/AndroidApp.kt
@@ -1,0 +1,28 @@
+package com.blackcrowsys
+
+import android.app.Activity
+import android.app.Application
+import dagger.android.AndroidInjector
+import dagger.android.DispatchingAndroidInjector
+import dagger.android.HasActivityInjector
+import com.blackcrowsys.di.DaggerAppComponent
+import com.facebook.drawee.backends.pipeline.Fresco
+import javax.inject.Inject
+
+class AndroidApp : Application(), HasActivityInjector {
+
+    @Inject lateinit var activityDispatchingAndroidInjector: DispatchingAndroidInjector<Activity>
+
+    override fun onCreate() {
+        super.onCreate()
+
+        DaggerAppComponent.builder()
+                .application(this)
+                .build()
+                .inject(this)
+
+        Fresco.initialize(this)
+    }
+
+    override fun activityInjector(): AndroidInjector<Activity> = activityDispatchingAndroidInjector
+}
