+++ app/src/main/java/com/blackcrowsys/api/model/IpAddress.kt
@@ -1,0 +1,3 @@
+package com.blackcrowsys.api.model
+
+data class IpAddress(val ip: String)
