+++ app/src/main/java/com/blackcrowsys/ui/main/di/MainActivityModule.kt
@@ -1,0 +1,6 @@
+package com.blackcrowsys.ui.main.di
+
+import dagger.Module
+
+@Module
+class MainActivityModule
