+++ app/src/main/java/com/blackcrowsys/util/SchedulerProvider.kt
@@ -1,0 +1,34 @@
+package com.blackcrowsys.util
+
+import io.reactivex.*
+
+class SchedulerProvider(val backgroundScheduler: Scheduler, val foregroundScheduler: Scheduler) {
+
+    fun <T> getSchedulersForObservable(): (Observable<T>) -> Observable<T> {
+        return { observable: Observable<T> ->
+            observable.subscribeOn(backgroundScheduler)
+                    .observeOn(foregroundScheduler)
+        }
+    }
+
+    fun <T> getSchedulersForSingle(): (Single<T>) -> Single<T> {
+        return { single: Single<T> ->
+            single.subscribeOn(backgroundScheduler)
+                    .observeOn(foregroundScheduler)
+        }
+    }
+
+    fun getSchedulersForCompletable(): (Completable) -> Completable {
+        return { completable: Completable ->
+            completable.subscribeOn(backgroundScheduler)
+                    .observeOn(foregroundScheduler)
+        }
+    }
+
+    fun <T> getSchedulersForFlowable(): (Flowable<T>) -> Flowable<T> {
+        return { flowable: Flowable<T> ->
+            flowable.subscribeOn(backgroundScheduler)
+                    .observeOn(foregroundScheduler)
+        }
+    }
+}
