--- app/src/main/java/uk/co/mycompany/AndroidApp.kt
@@ -1,25 +1,0 @@
-package uk.co.mycompany
-
-import android.app.Activity
-import android.app.Application
-import dagger.android.AndroidInjector
-import dagger.android.DispatchingAndroidInjector
-import dagger.android.HasActivityInjector
-import uk.co.mycompany.di.DaggerAppComponent
-import javax.inject.Inject
-
-class AndroidApp: Application(), HasActivityInjector {
-
-    @Inject lateinit var activityDispatchingAndroidInjector: DispatchingAndroidInjector<Activity>
-
-    override fun onCreate() {
-        super.onCreate()
-
-        DaggerAppComponent.builder()
-                .application(this)
-                .build()
-                .inject(this)
-    }
-
-    override fun activityInjector(): AndroidInjector<Activity> = activityDispatchingAndroidInjector
-}
