--- app/src/main/java/uk/co/mycompany/api/ApiService.kt
@@ -1,11 +1,0 @@
-package uk.co.mycompany.api
-
-import io.reactivex.Single
-import retrofit2.http.GET
-import uk.co.mycompany.api.model.IpAddress
-
-interface ApiService {
-
-    @GET(".")
-    fun getJsonResponse(): Single<IpAddress>
-}
