--- app/src/main/java/uk/co/mycompany/api/model/IpAddress.kt
@@ -1,3 +1,0 @@
-package uk.co.mycompany.api.model
-
-data class IpAddress(val ip: String)
