--- app/src/main/java/uk/co/mycompany/di/ActivityBuilder.kt
@@ -1,13 +1,0 @@
-package uk.co.mycompany.di
-
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-import uk.co.mycompany.ui.main.MainActivity
-import uk.co.mycompany.ui.main.di.MainActivityModule
-
-@Module
-abstract class ActivityBuilder {
-
-    @ContributesAndroidInjector(modules = arrayOf(MainActivityModule::class))
-    abstract fun bindMainActivity(): MainActivity
-}
