--- app/src/main/java/uk/co/mycompany/di/AppModule.kt
@@ -1,67 +1,0 @@
-package uk.co.mycompany.di
-
-import android.app.Application
-import com.google.gson.FieldNamingPolicy
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import dagger.Module
-import dagger.Provides
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-import okhttp3.Cache
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import uk.co.mycompany.api.ApiService
-import uk.co.mycompany.util.SchedulerProvider
-import java.io.File
-import java.util.*
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@Module
-class AppModule {
-
-    @Provides
-    @Singleton
-    fun provideSchedulerProvider() = SchedulerProvider(Schedulers.io(), AndroidSchedulers.mainThread())
-
-    @Provides
-    @Singleton
-    fun provideGson(): Gson {
-        return GsonBuilder().setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-                .create()
-    }
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(application: Application): OkHttpClient {
-        val interceptor = HttpLoggingInterceptor()
-        interceptor.level = HttpLoggingInterceptor.Level.BASIC
-
-        val cacheDir = File(application.cacheDir, UUID.randomUUID().toString())
-        // 10 MiB cache
-        val cache = Cache(cacheDir, 10 * 1024 * 1024)
-
-        return OkHttpClient.Builder()
-                .cache(cache)
-                .connectTimeout(30, TimeUnit.SECONDS)
-                .readTimeout(60, TimeUnit.SECONDS)
-                .writeTimeout(60, TimeUnit.SECONDS)
-                .addInterceptor(interceptor)
-                .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideApiService(gson: Gson, okHttpClient: OkHttpClient): ApiService {
-        return Retrofit.Builder()
-                .baseUrl("http://ip.jsontest.com/")
-                .addConverterFactory(GsonConverterFactory.create(gson))
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .client(okHttpClient)
-                .build().create(ApiService::class.java)
-    }
-}
