--- app/src/main/java/uk/co/mycompany/util/SchedulerProvider.kt
@@ -1,34 +1,0 @@
-package uk.co.mycompany.util
-
-import io.reactivex.*
-
-class SchedulerProvider(val backgroundScheduler: Scheduler, val foregroundScheduler: Scheduler) {
-
-    fun <T> getSchedulersForObservable(): (Observable<T>) -> Observable<T> {
-        return { observable: Observable<T> ->
-            observable.subscribeOn(backgroundScheduler)
-                    .observeOn(foregroundScheduler)
-        }
-    }
-
-    fun <T> getSchedulersForSingle(): (Single<T>) -> Single<T> {
-        return { single: Single<T> ->
-            single.subscribeOn(backgroundScheduler)
-                    .observeOn(foregroundScheduler)
-        }
-    }
-
-    fun getSchedulersForCompletable(): (Completable) -> Completable {
-        return { completable: Completable ->
-            completable.subscribeOn(backgroundScheduler)
-                    .observeOn(foregroundScheduler)
-        }
-    }
-
-    fun <T> getSchedulersForFlowable(): (Flowable<T>) -> Flowable<T> {
-        return { flowable: Flowable<T> ->
-            flowable.subscribeOn(backgroundScheduler)
-                    .observeOn(foregroundScheduler)
-        }
-    }
-}
