+++ app/src/main/res/drawable/splash_background.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:gravity="fill"
+        android:drawable="@color/white" />
+    <item>
+        <bitmap
+            android:gravity="center"
+            android:src="@mipmap/dinewell_logo_blue_green" />
+    </item>
+
+</layer-list>
