+++ build.gradle
@@ -31,7 +31,7 @@ ext {
     targetSdkVersion = 26
     compileSdkVersion = 26
     buildToolsVersion = "26.0.2"
-    supportLibraryVersion = "27.0.1"
+    supportLibraryVersion = "27.0.2"
 
     architectureComponentsVersion = "1.0.0"
     daggerVersion = "2.13"
